updated template for CentOS 7.7

+++ docker_hosts/centos/scripts/update.sh
@@ -1,5 +1,12 @@
-#!/bin/sh -eux
+#!/bin/bash -eux
 
-yum -y update;
-reboot;
-sleep 60;
+if [[ $UPDATE  =~ true || $UPDATE =~ 1 || $UPDATE =~ yes ]]; then
+    yum -y -q update
+    echo "==> Updates applied, rebooting..."
+
+    # reboot
+    reboot
+    sleep 60
+fi
+
+echo "==> update.sh skipped"
